requirements.txt
pyrogram==2.0.106
python-dotenv==1.0.1
requests==2.31.0
pillow==10.3.0
pytesseract==0.3.10
